fixed all db bugs and user authorizations

+++ app/Http/Controllers/AboutsController.php
@@ -16,9 +16,13 @@ class AboutsController extends Controller
      */
     public function index()
     {
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        $abouts = About::get();
-        return view("admin.abouts.index")->with("abouts", $abouts);
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            $abouts = About::get();
+            return view("admin.abouts.index")->with("abouts", $abouts);
+        } else {
+            return redirect('/login');
+        }
     }
 
     /**
@@ -28,15 +32,19 @@ class AboutsController extends Controller
      */
     public function create()
     {
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        $about = About::get();
-        if(count($about) > 0) {
-            Session::flash("info", "An About page already exists edit it instead");
-            return redirect("/admin/abouts");
-        }
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            $about = About::get();
+            if(count($about) > 0) {
+                Session::flash("info", "An About page already exists edit it instead");
+                return redirect("/admin/abouts");
+            }
 
-        $mode = ["edit" => false, "add" => true];
-        return view("admin.abouts.form")->with("mode", $mode);
+            $mode = ["edit" => false, "add" => true];
+            return view("admin.abouts.form")->with("mode", $mode);
+        } else {
+            return redirect('/login');
+        }
     }
 
     /**
@@ -47,27 +55,31 @@ class AboutsController extends Controller
      */
     public function store(Request $request, About $about)
     {
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        $data = [];
-        if($request->isMethod("POST"))
-        {
-            $this->validate($request,
-                [
-                    "photo" => "required|mimes:jpeg,bmp,jpg,png,gif",
-                    "content" => "required"
-                ]
-            );
-            if($request->hasfile("photo")) 
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            $data = [];
+            if($request->isMethod("POST"))
             {
-                $photo = $request->photo;
-                $name = time().$photo->getClientOriginalName();
-                $photo->move(public_path("/images/abouts"), $name);
-                $data["photo"] = $name;
+                $this->validate($request,
+                    [
+                        "photo" => "required|mimes:jpeg,bmp,jpg,png,gif",
+                        "content" => "required"
+                    ]
+                );
+                if($request->hasfile("photo")) 
+                {
+                    $photo = $request->photo;
+                    $name = time().$photo->getClientOriginalName();
+                    $photo->move(public_path("/images/abouts"), $name);
+                    $data["photo"] = $name;
+                }
+                $data["content"] = $request->content;
+                $about->insert($data);
+                Session::flash("success", "About page added successfully");
+                return redirect("/admin/abouts");
             }
-            $data["content"] = $request->content;
-            $about->insert($data);
-            Session::flash("success", "About page added successfully");
-            return redirect("/admin/abouts");
+        } else {
+            return redirect('/login');
         }
     }
 
@@ -79,8 +91,12 @@ class AboutsController extends Controller
      */
     public function show($id)
     {
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        return redirect("/admin/abouts");
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            return redirect("/admin/abouts");
+        } else {
+            return redirect("/login");
+        }
     }
 
     /**
@@ -91,10 +107,14 @@ class AboutsController extends Controller
      */
     public function edit($id)
     {
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        $about = About::find($id);
-        $mode = ["edit" => true, "add" => false];
-        return view("admin.abouts.form")->with(["mode" => $mode, "about" => $about]);
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            $about = About::find($id);
+            $mode = ["edit" => true, "add" => false];
+            return view("admin.abouts.form")->with(["mode" => $mode, "about" => $about]);
+        } else {
+            return redirect('/login');
+        }
     }
 
     /**
@@ -106,28 +126,32 @@ class AboutsController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(Auth::user()) {
         if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
         $about = About::find($id);
-        if($request->isMethod("PUT"))
-        {
-            $this->validate(
-                $request,
-                [
-                    "photo" => "mimes:jpeg,png,bmp,gif",
-                    "content" => "required"
-                ]
-            );
-            if($request->photo) 
+            if($request->isMethod("PUT"))
             {
-                $photo = $request->photo;
-                $name = time().$photo->getClientOriginalName();
-                $photo->move(public_path("/images/abouts"), $name);
-                $data["photo"] = $name;
+                $this->validate(
+                    $request,
+                    [
+                        "photo" => "mimes:jpeg,png,bmp,gif",
+                        "content" => "required"
+                    ]
+                );
+                if($request->photo) 
+                {
+                    $photo = $request->photo;
+                    $name = time().$photo->getClientOriginalName();
+                    $photo->move(public_path("/images/abouts"), $name);
+                    $data["photo"] = $name;
+                }
+                $data["content"] = $request->content;
+                $about->update($data);
+                Session::flash("success", "About page updated successfully");
+                return redirect("/admin/abouts");
             }
-            $data["content"] = $request->content;
-            $about->update($data);
-            Session::flash("success", "About page updated successfully");
-            return redirect("/admin/abouts");
+        } else {
+            return redirect('/login');
         }
     }
 
@@ -139,10 +163,14 @@ class AboutsController extends Controller
      */
     public function destroy($id)
     {
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        $about = About::find($id);
-        $about->delete();
-        Session::flash("success", "About was deleted successfully");
-        return redirect("/admin/abouts");
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            $about = About::find($id);
+            $about->delete();
+            Session::flash("success", "About was deleted successfully");
+            return redirect("/admin/abouts");
+        } else {
+            return redirect('/login');
+        }
     }
 }
