+++ app/Http/Controllers/ProductsController.php
@@ -17,12 +17,17 @@ class ProductsController extends Controller
      */
     public function index(Product $product)
     {
-        $products = Product::get();
-        $categories = Category::get();
-        foreach ($products as $product) {
-            $product->images = json_decode($product->images);
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+                $products = Product::get();
+                $categories = Category::get();
+                foreach ($products as $product) {
+                    $product->images = json_decode($product->images);
+                }
+                return view("admin.products.index")->with(["products" =>$products, "categories" => $categories]);
+        } else {
+            return redirect('/login');
         }
-        return view("admin.products.index")->with(["products" =>$products, "categories" => $categories]);
     }
 
     /**
@@ -32,8 +37,13 @@ class ProductsController extends Controller
      */
     public function create()
     {
-        $categories  = Category::get();
-        return view("admin.products.form")->with(["categories" => $categories]);
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+                $categories  = Category::get();
+                return view("admin.products.form")->with(["categories" => $categories]);
+        } else {
+            return redirect('/login');
+        }
     }
 
     /**
@@ -44,39 +54,44 @@ class ProductsController extends Controller
      */
     public function store(Request $request, Product $product)
     {
-        $data = [];
-        $data["name"] = $request->name;
-        $data["price"] = $request->price;
-        $data["category_id"] = $request->category;
-        $data["description"] = $request->description;
-        $data["status"] = $request->status;
-        if($request->isMethod("POST")) {
-            $this->validate(
-                $request,
-                [
-                    "name" => "required",
-                    "price" => "required",
-                    "images.*" => "required|mimes:jpeg,png,bmp,gif",
-                    "category" => "required",
-                    "description" => "required",
-                    "status" => "required" 
-                ]
-            );
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+                $data = [];
+                $data["name"] = $request->name;
+                $data["price"] = $request->price;
+                $data["category_id"] = $request->category;
+                $data["description"] = $request->description;
+                $data["status"] = $request->status;
+                if($request->isMethod("POST")) {
+                    $this->validate(
+                        $request,
+                        [
+                            "name" => "required",
+                            "price" => "required",
+                            "images.*" => "required|mimes:jpeg,png,bmp,gif",
+                            "category" => "required",
+                            "description" => "required",
+                            "status" => "required" 
+                        ]
+                    );
 
-            if($request->hasfile("images")) 
-            {
-                $image_names = [];
-                $images_files = $request->images;
-                foreach ($images_files as $image) {
-                    $name = time().$image->getClientOriginalName();
-                    $image->move(public_path("/images/products"), $name);
-                    $image_names[] = $name;
+                    if($request->hasfile("images")) 
+                    {
+                        $image_names = [];
+                        $images_files = $request->images;
+                        foreach ($images_files as $image) {
+                            $name = time().$image->getClientOriginalName();
+                            $image->move(public_path("/images/products"), $name);
+                            $image_names[] = $name;
+                        }
+                    }
+                    $data["images"] = json_encode($image_names);
+                    $product->insert($data);
+                    Session::flash("success", "Product was added successfully");
+                    return redirect("/admin/products");
                 }
-            }
-            $data["images"] = json_encode($image_names);
-            $product->insert($data);
-            Session::flash("success", "Product was added successfully");
-            return redirect("/admin/products");
+        } else {
+            return redirect('/login');
         }
     }
 
@@ -88,11 +103,16 @@ class ProductsController extends Controller
      */
     public function show($id)
     {
-        $product = Product::findOrFail($id);
-        $product->images = json_decode($product->images);
-        $category = Category::findOrFail($product->id);
-        return view("admin.products.product")
-                ->with(["product"=>$product, "category"=>$category]);
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+                $product = Product::findOrFail($id);
+                $product->images = json_decode($product->images);
+                $category = Category::findOrFail($product->id);
+                return view("admin.products.product")
+                        ->with(["product"=>$product, "category"=>$category]);
+        } else {
+            return redirect('/login');
+        }
     }
 
     /**
@@ -103,10 +123,15 @@ class ProductsController extends Controller
      */
     public function edit($id)
     {
-        $categories = Category::get();
-        $product = Product::findOrFail($id);
-        return view("admin.products.edit-form")
-        ->with(["categories" => $categories, "product" => $product]);
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+                $categories = Category::get();
+                $product = Product::findOrFail($id);
+                return view("admin.products.edit-form")
+                ->with(["categories" => $categories, "product" => $product]);
+        } else {
+            return redirect('login');
+        }
     }
 
     /**
@@ -118,47 +143,52 @@ class ProductsController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $product = Product::findOrFail($id);
-   
-        $data = [];
-        $data["name"] = $request->name;
-        $data["price"] = $request->price;
-        $data["category_id"] = $request->category;
-        $data["description"] = $request->description;
-        $data["status"] = $request->status;
-        if($request->isMethod("PUT")) {
-            $this->validate(
-                $request,
-                [
-                    "name" => "required",
-                    "price" => "required",
-                    "images.*" => "required|mimes:jpeg,png,bmp,gif",
-                    "category" => "required",
-                    "description" => "required",
-                    "status" => "required" 
-                ]
-            );
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+                $product = Product::findOrFail($id);
+        
+                $data = [];
+                $data["name"] = $request->name;
+                $data["price"] = $request->price;
+                $data["category_id"] = $request->category;
+                $data["description"] = $request->description;
+                $data["status"] = $request->status;
+                if($request->isMethod("PUT")) {
+                    $this->validate(
+                        $request,
+                        [
+                            "name" => "required",
+                            "price" => "required",
+                            "images.*" => "required|mimes:jpeg,png,bmp,gif",
+                            "category" => "required",
+                            "description" => "required",
+                            "status" => "required" 
+                        ]
+                    );
 
-            if($request->hasfile("images")) 
-            {
-                $image_names = [];
-                $images_files = $request->images;
-                foreach ($images_files as $image) {
-                    $name = time().$image->getClientOriginalName();
-                    $image->move(public_path("/images/products"), $name);
-                    $image_names[] = $name;
-                    $data["images"] = json_encode($image_names);
-                }
-                foreach(json_decode($product->images) as $image) {
-                    $file = public_path("/images/products/").$image;
-                    if(file_exists($file)) {
-                        unlink($file);
+                    if($request->hasfile("images")) 
+                    {
+                        $image_names = [];
+                        $images_files = $request->images;
+                        foreach ($images_files as $image) {
+                            $name = time().$image->getClientOriginalName();
+                            $image->move(public_path("/images/products"), $name);
+                            $image_names[] = $name;
+                            $data["images"] = json_encode($image_names);
+                        }
+                        foreach(json_decode($product->images) as $image) {
+                            $file = public_path("/images/products/").$image;
+                            if(file_exists($file)) {
+                                unlink($file);
+                            }
+                        }
                     }
+                    Session::flash("success", "Product was update successful");
+                    $product->update($data);
+                    return redirect("/admin/products");
                 }
-            }
-            Session::flash("success", "Product was update successful");
-            $product->update($data);
-            return redirect("/admin/products");
+        } else {
+            return redirect('/login');
         }
     }
 
@@ -170,31 +200,41 @@ class ProductsController extends Controller
      */
     public function destroy($id)
     {
-        $product = Product::findOrFail($id);
-        foreach(json_decode($product->images) as $image) {
-            $file = public_path("images/products/").$image;
-            if(file_exists($file)) {
-                unlink($file);
-            }
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+                $product = Product::findOrFail($id);
+                foreach(json_decode($product->images) as $image) {
+                    $file = public_path("images/products/").$image;
+                    if(file_exists($file)) {
+                        unlink($file);
+                    }
+                }
+                $product->delete();
+                Session::flash("success", "Product was deleted successfully");
+                return redirect("/admin/products");
+        } else {
+            return redirect('/login');
         }
-        $product->delete();
-        Session::flash("success", "Product was deleted successfully");
-        return redirect("/admin/products");
         
     }
 
     public function search(Request $request) 
     {
-        $products = Product::where("name", "like", "%". request("query") . "%")->take(5)->get();
-        foreach($products as $product) {
-            $product->images = json_decode($product->images);
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+                $products = Product::where("name", "like", "%". request("query") . "%")->take(5)->get();
+                foreach($products as $product) {
+                    $product->images = json_decode($product->images);
+                }
+                if(request("ajax")) {
+                    return response()->json($products, 200);
+                }
+                return view("admin.products.search")->with([
+                    "products" => $products, 
+                    "query"=>request("query")
+                    ]);
+        } else {
+            return redirect('/login');
         }
-        if(request("ajax")) {
-            return response()->json($products, 200);
-        }
-        return view("admin.products.search")->with([
-            "products" => $products, 
-            "query"=>request("query")
-            ]);
     }
 }
