+++ app/Http/Controllers/SettingsController.php
@@ -16,9 +16,13 @@ class SettingsController extends Controller
      */
     public function index()
     {
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        $settings = Setting::get();
-        return view("admin.settings.index")->with("settings", $settings);
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            $settings = Setting::get();
+            return view("admin.settings.index")->with("settings", $settings);
+        } else {
+            return redirect('/login');
+        }
     }
 
     /**
@@ -28,13 +32,17 @@ class SettingsController extends Controller
      */
     public function create()
     {
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        $settings = Setting::get();
-        if(count($settings) > 0) {
-            Session::flash("info", "A Setting already exists edit it instead");
-            return redirect("/admin/settings");
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            $settings = Setting::get();
+            if(count($settings) > 0) {
+                Session::flash("info", "A Setting already exists edit it instead");
+                return redirect("/admin/settings");
+            }
+            return view("admin.settings.form");
+        } else {
+            return redirect('login');
         }
-        return view("admin.settings.form");
     }
 
     /**
@@ -45,20 +53,24 @@ class SettingsController extends Controller
      */
     public function store(Request $request)
     {
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        $data = [];
-        $data["facebook"] = $request->facebook;
-        $data["instagram"] = $request->instagram;
-        $data["twitter"] = $request->twitter;
-        $data["whatsapp"] = $request->whatsapp;
-        $data["phone_1"] = $request->phone_1;
-        $data["phone_2"] = $request->phone_2;
-        $data["email_1"] = $request->email_1;
-        $data["email_2"] = $request->email_2;
-        $data["address"] = $request->address;
-        Setting::create($data);
-        Session::flash("success", "Setting was added successfully");
-        return redirect("/admin/settings");
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            $data = [];
+            $data["facebook"] = $request->facebook;
+            $data["instagram"] = $request->instagram;
+            $data["twitter"] = $request->twitter;
+            $data["whatsapp"] = $request->whatsapp;
+            $data["phone_1"] = $request->phone_1;
+            $data["phone_2"] = $request->phone_2;
+            $data["email_1"] = $request->email_1;
+            $data["email_2"] = $request->email_2;
+            $data["address"] = $request->address;
+            Setting::create($data);
+            Session::flash("success", "Setting was added successfully");
+            return redirect("/admin/settings");
+        } else {
+            return redirect('/login');
+        }
     }
 
     /**
@@ -69,8 +81,12 @@ class SettingsController extends Controller
      */
     public function show($id)
     {
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        return redirect("/admin/settings");
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            return redirect("/admin/settings");
+        } else {
+            return redirect('/login');
+        }
     }
 
     /**
@@ -81,9 +97,13 @@ class SettingsController extends Controller
      */
     public function edit($id)
     {
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        $setting = Setting::findOrFail($id);
-        return view("admin.settings.edit-form")->with("setting", $setting);
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            $setting = Setting::findOrFail($id);
+            return view("admin.settings.edit-form")->with("setting", $setting);
+        } else {
+            return redirect('/login');
+        }
     }
 
     /**
@@ -95,22 +115,26 @@ class SettingsController extends Controller
      */
     public function update(Request $request, $id)
     {   
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        $setting = Setting::findOrFail($id);
-        $data = [];
-        $data["facebook"] = $request->facebook;
-        $data["instagram"] = $request->instagram;
-        $data["twitter"] = $request->twitter;
-        $data["whatsapp"] = $request->whatsapp;
-        $data["phone_1"] = $request->phone_1;
-        $data["phone_2"] = $request->phone_2;
-        $data["email_1"] = $request->email_1;
-        $data["email_2"] = $request->email_2;
-        $data["address"] = $request->address;
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            $setting = Setting::findOrFail($id);
+            $data = [];
+            $data["facebook"] = $request->facebook;
+            $data["instagram"] = $request->instagram;
+            $data["twitter"] = $request->twitter;
+            $data["whatsapp"] = $request->whatsapp;
+            $data["phone_1"] = $request->phone_1;
+            $data["phone_2"] = $request->phone_2;
+            $data["email_1"] = $request->email_1;
+            $data["email_2"] = $request->email_2;
+            $data["address"] = $request->address;
 
-        $setting->update($data);
-        Session::flash("success", "Setting was updated successfully");
-        return redirect("/admin/settings");
+            $setting->update($data);
+            Session::flash("success", "Setting was updated successfully");
+            return redirect("/admin/settings");
+        } else {
+            return redirect('/login');
+        }
     }
 
     /**
@@ -121,11 +145,14 @@ class SettingsController extends Controller
      */
     public function destroy($id)
     {
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        $setting = Setting::findOrFail($id);
-        $setting->delete();
-        Session::flash("success", "Setting was deleted successfully");
-        return redirect("/admin/settings");
-        
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            $setting = Setting::findOrFail($id);
+            $setting->delete();
+            Session::flash("success", "Setting was deleted successfully");
+            return redirect("/admin/settings");
+        } else {
+            return redirect('/login');
+        }
     }
 }
