+++ app/Http/Controllers/SlidesController.php
@@ -16,9 +16,13 @@ class SlidesController extends Controller
      */
     public function index()
     {
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        $slides = Slides::get();
-        return view("admin.slides.index")->with("slides", $slides);
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            $slides = Slides::get();
+            return view("admin.slides.index")->with("slides", $slides);
+        } else {
+            return redirect("/login");
+        }
     }
 
     /**
@@ -28,9 +32,13 @@ class SlidesController extends Controller
      */
     public function create()
     {
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        $mode = ["edit" => false, "add" => true];
-        return view("admin.slides.form")->with("mode", $mode);
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            $mode = ["edit" => false, "add" => true];
+            return view("admin.slides.form")->with("mode", $mode);
+        } else {
+            return redirect('/login');
+        }
     }
 
     /**
@@ -41,27 +49,31 @@ class SlidesController extends Controller
      */
     public function store(Request $request)
     {
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        $data = [];
-        if($request->isMethod("POST")) {
-            $this->validate($request, 
-                [
-                    "title" => "required",
-                    "photo" => "required|mimes:png,jpeg,gif,bmp",
-                    "info" => "required|max:250"
-                ]
-            );
-            if($request->hasfile("photo")) 
-            {
-                $name = time().$request->photo->getClientOriginalName();
-                $data["photo"] = $name;
-                $request->photo->move(public_path("/images/slides"), $name);
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            $data = [];
+            if($request->isMethod("POST")) {
+                $this->validate($request, 
+                    [
+                        "title" => "required",
+                        "photo" => "required|mimes:png,jpeg,gif,bmp",
+                        "info" => "required|max:250"
+                    ]
+                );
+                if($request->hasfile("photo")) 
+                {
+                    $name = time().$request->photo->getClientOriginalName();
+                    $data["photo"] = $name;
+                    $request->photo->move(public_path("/images/slides"), $name);
+                }
+                $data["title"] = $request->title;
+                $data["info"] = $request->info;
+                Slides::create($data);
+                Session::flash("success", "Slide was added successfully");
+                return redirect("/admin/slides");
             }
-            $data["title"] = $request->title;
-            $data["info"] = $request->info;
-            Slides::create($data);
-            Session::flash("success", "Slide was added successfully");
-            return redirect("/admin/slides");
+        } else {
+            return redirect('/login');
         }
     }
 
@@ -73,8 +85,12 @@ class SlidesController extends Controller
      */
     public function show($id)
     {
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        return redirect("/admin/settings");
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            return redirect("/admin/settings");
+        } else {
+            return redirect('/login');
+        }
     }
 
     /**
@@ -84,10 +100,14 @@ class SlidesController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
-    {
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        $slide = Slides::find($id);
-        return view("admin.slides.edit-form")->with(["slide" => $slide]);
+    {   
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            $slide = Slides::find($id);
+            return view("admin.slides.edit-form")->with(["slide" => $slide]);
+        } else {
+            return redirect('/login');
+        }
     }
 
     /**
@@ -99,34 +119,38 @@ class SlidesController extends Controller
      */
     public function update(Request $request, $id)
     {
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        $slide = Slides::find($id);
-        $data = [];
-        if($request->isMethod("PUT"))
-        {
-            $this->validate($request, 
-                [
-                    "title" => "required",
-                    "photo" => "mimes:jpeg,png,bmp,gif",
-                    "info" => "required|max:250"
-                ]
-            );
-            if($request->hasfile("photo"))
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            $slide = Slides::find($id);
+            $data = [];
+            if($request->isMethod("PUT"))
             {
-                $file = public_path("/images/slides/").$slide->photo;
-                if(file_exists($file)) 
+                $this->validate($request, 
+                    [
+                        "title" => "required",
+                        "photo" => "mimes:jpeg,png,bmp,gif",
+                        "info" => "required|max:250"
+                    ]
+                );
+                if($request->hasfile("photo"))
                 {
-                    unlink($file);
+                    $file = public_path("/images/slides/").$slide->photo;
+                    if(file_exists($file)) 
+                    {
+                        unlink($file);
+                    }
+                    $name = time().$request->photo->getClientOriginalName();
+                    $request->photo->move(public_path("/images/slides"), $name);
+                    $data["photo"] = $name;
                 }
-                $name = time().$request->photo->getClientOriginalName();
-                $request->photo->move(public_path("/images/slides"), $name);
-                $data["photo"] = $name;
+                $data["title"] = $request->title;
+                $data["info"] = $request->info;
+                $slide->update($data);
+                Session::flash("success", "Slide was updated successfully");
+                return redirect("/admin/slides");
             }
-            $data["title"] = $request->title;
-            $data["info"] = $request->info;
-            $slide->update($data);
-            Session::flash("success", "Slide was updated successfully");
-            return redirect("/admin/slides");
+        } else {
+            return redirect('/login');
         }
     }
 
@@ -138,17 +162,21 @@ class SlidesController extends Controller
      */
     public function destroy($id)
     {
-        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
-        $slide = Slides::find($id);
-        if($slide)
-        {
-            $file = public_path("images/slides/").$slide->photo;
-            if(file_exists($file)) {
-                unlink($file);
+        if(Auth::user()) {
+            if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+            $slide = Slides::find($id);
+            if($slide)
+            {
+                $file = public_path("images/slides/").$slide->photo;
+                if(file_exists($file)) {
+                    unlink($file);
+                }
+                $slide->delete();
+                Session::flash("success", "Slide was deleted successfully");
+                return redirect("/admin/slides");
             }
-            $slide->delete();
-            Session::flash("success", "Slide was deleted successfully");
-            return redirect("/admin/slides");
+        } else {
+            return redirect('/login');
         }
     }
 }
